requirements.txt
Flask
firebase
requests_toolbelt
sseclient
gcloud
python_jwt